wanderlust fixed urneachable estimation and more
do a BFS to attach unreachable populations in wanderlust algorithm.
also fixed 'plot as function' for unnormalized wanderlust channel and
added an option to stretch dens area in the plot.
fixed subsample to account to intersect mode
added ability to create gate from cluster channel that will hold the
means
fixed a bug in plotScatter